third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-canvas-tainting-video-with-range-request.https-expected.txt

This is a testharness.js-based test.
PASS range responses from single origin (same-origin)
FAIL range responses from multiple origins (cross-origin first) assert_equals: expected "TAINTED" but got "LOAD_ERROR"
Harness: the test ran to completion.
